split world_builder off into its own class
