+++ Vic3ToHoI4lib.vcxproj
@@ -146,6 +146,7 @@
     <ClInclude Include="src\vic3_world\states\vic3_states_importer.h" />
     <ClInclude Include="src\vic3_world\states\vic3_state_importer.h" />
     <ClInclude Include="src\vic3_world\world\vic3_world.h" />
+    <ClInclude Include="src\vic3_world\world\vic3_world_builder.h" />
     <ClInclude Include="src\vic3_world\world\vic3_world_importer.h" />
   </ItemGroup>
   <ItemGroup>
@@ -250,6 +251,7 @@
     <ClCompile Include="src\vic3_world\technology\vic3_technology_importer.cpp" />
     <ClInclude Include="src\vic3_world\technology\vic3_technology_importer.h" />
     <ClCompile Include="src\vic3_world\world\vic3_world.cpp" />
+    <ClCompile Include="src\vic3_world\world\vic3_world_builder.cpp" />
     <ClCompile Include="src\vic3_world\world\vic3_world_importer.cpp" />
   </ItemGroup>
   <ItemGroup>
