+++ Vic3ToHoI4lib.vcxproj.filters
@@ -573,6 +573,9 @@
     <ClInclude Include="src\mappers\world\world_mapper_builder.h">
       <Filter>src\mappers\world</Filter>
     </ClInclude>
+    <ClInclude Include="src\vic3_world\world\vic3_world_builder.h">
+      <Filter>src\vic3_world\world</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\vic3_to_hoi4_converter.cpp">
@@ -877,6 +880,9 @@
     </ClCompile>
     <ClCompile Include="src\mappers\world\world_mapper_builder.cpp">
       <Filter>src\mappers\world</Filter>
+    </ClCompile>
+    <ClCompile Include="src\vic3_world\world\vic3_world_builder.cpp">
+      <Filter>src\vic3_world\world</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
